added cart and user model

+++ src/main/java/com/ugurukku/kitabal/entities/Book.java
@@ -30,7 +30,7 @@ public class Book {
 
     Integer page;
 
-    Integer rate;
+    Double rate;
 
     Integer stock;
